projects/coinflip-game-contracts/pyproject.toml
[tool.poetry]
name = "coinflip-game-contracts"
version = "0.1.0"
description = "Algorand smart contracts"
authors = ["VyNguyen <vynguyenj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
beaker-pyteal = "^1.1.1"
algokit-utils = "^2.2.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
setuptools = "^69.0.2" # Adding explicitly to work around pyteal https://github.com/algorand/pyteal/issues/712

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
